backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
celery==5.2.7
redis==4.5.5
beautifulsoup4==4.11.2
requests==2.30.0
lxml==5.3.0
playwright==1.48.0
weasyprint==59.0
python-dotenv==1.0.0
Pillow==11.0.0
